pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversational-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "jupyter>=1.1.1",
    "notebook>=7.5.5",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langgraph>=1.1.8",
    "langchain-community>=0.4.1",
    "google-genai>=1.73.1",
    "langchain-google-genai>=4.2.2",
    # "langchain-google-vertexai>=3.2.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.49",
    "faker>=20.0.0",
    "redis>=7.4.0",
    "langgraph-checkpoint-redis>=0.4.1",
    "pydantic-settings>=2.13.1",
    "fastapi>=0.136.0",
    "uvicorn>=0.44.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
